[chore] health entity로 통일
[chore] health entity로 통일
